refactor: delete H1 from TITLEs strings
